Rename matchers pkg to accuracy
