+++ validator/src/main/java/com/github/maracas/accuracy/LocationMatcher.java
@@ -1,4 +1,4 @@
-package com.github.maracas.matchers;
+package com.github.maracas.accuracy;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -8,9 +8,9 @@ import java.util.Map;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import com.github.maracas.accuracy.AccuracyCase.AccuracyType;
 import com.github.maracas.brokenUse.BrokenUse;
 import com.github.maracas.build.CompilerMessage;
-import com.github.maracas.matchers.AccuracyCase.AccuracyType;
 
 /**
  * Type of matcher based on {@link BrokenUse} and {@link CompilerMessage}
@@ -25,7 +25,7 @@ public class LocationMatcher implements Matcher {
     @Override
     public Collection<AccuracyCase> match(Collection<BrokenUse> brokenUses, List<CompilerMessage> messages) {
         Map<String, List<CompilerMessage>> messagesMap = Matcher.messagesToMap(messages);
-        Collection<AccuracyCase> matches = new ArrayList<AccuracyCase>();
+        Collection<AccuracyCase> cases = new ArrayList<AccuracyCase>();
 
         for (BrokenUse brokenUse : brokenUses) {
             String path = brokenUse.element().getPosition().getFile().getAbsolutePath();
@@ -43,14 +43,14 @@ public class LocationMatcher implements Matcher {
                 }
 
             if (!matchedMessages.isEmpty()) {
-                matches.add(new AccuracyCase(brokenUse, matchedMessages, AccuracyType.TRUE_POSITIVE));
+                cases.add(new AccuracyCase(brokenUse, matchedMessages, AccuracyType.TRUE_POSITIVE));
                 System.out.println(AccuracyType.TRUE_POSITIVE);
             } else {
-                matches.add(new AccuracyCase(brokenUse, null, AccuracyType.FALSE_POSITIVE));
+                cases.add(new AccuracyCase(brokenUse, null, AccuracyType.FALSE_POSITIVE));
                 System.out.println(AccuracyType.FALSE_POSITIVE);
             }
 
         }
-        return matches;
+        return cases;
     }
 }
